auth-flask/requirements.txt
cffi==1.15.1
click==8.1.3
colorama==0.4.6
cryptography==39.0.0
Flask==2.2.2
Flask-JWT-Extended==4.4.4
Flask-MySQL==1.5.2
Flask-SQLAlchemy==3.0.2
greenlet==2.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pycparser==2.21
PyJWT==2.6.0
PyMySQL==1.0.2
SQLAlchemy==1.4.46
Werkzeug==2.2.2
